requirements.txt
Flask==2.1.2
gekko==1.0.4
numpy==1.23.0
PyYAML==6.0
requests==2.27.1
pycryptodome==3.15.0